requirements.txt
pandas
numpy

matplotlib
seaborn

requests
python-dotenv
ipykernel